requirements.txt
Flask>=2.0.0
flask-cors>=3.0.10
google-cloud-translate>=3.12.0
google-cloud-texttospeech>=2.14.1
google-cloud-speech>=2.21.0
pydub>=0.25.1
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.20.3
gunicorn>=20.0.0
blip
ffmpeg